Add another batch of refactoring to answer some review comments
