+++ packages/snaps-execution-environments/src/common/endowments/hardenedEndowments.ava.test.ts
@@ -5,10 +5,8 @@ import test from 'ava';
 // eslint-disable-next-line import/no-extraneous-dependencies, @typescript-eslint/no-unused-vars
 import FinalizationRegistry from 'globals';
 
-import {
-  testEndowmentHardening,
-  walkAndSearch,
-} from '../test-utils/endowments';
+import { walkAndSearch } from '../test-utils/endowments';
+import { testEndowmentHardening } from '../test-utils/hardening';
 import buildCommonEndowments from './commonEndowmentFactory';
 import Crypto from './crypto';
 import date from './date';
